feat(dbx): support mysql and sqlite alongside postgres
dbx.New dispatches by cfg.Driver (postgres|mysql|sqlite) to the matching
GORM dialector. Config holds per-dialect pointer sub-configs
(Postgres/MySQL/SQLite) plus shared pool/GORM options, so callers switch
databases by changing config only. SQLite uses the pure-Go
github.com/glebarez/sqlite (no CGO); MySQL DSN is built via
go-sql-driver/mysql Config. SetupTestDB(t, driver) covers all three
(sqlite via temp file, postgres/mysql via testcontainers).

PostgresConfig gains Schema (connection-level search_path).

README/CLAUDE.md updated for the multi-dialect API.

+++ dbx/pagination.go
@@ -74,7 +74,7 @@ func OffsetPaginate[T any](tx *gorm.DB, p PageParams) (*PageResult[T], error) {
 	var list []T
 	if !p.Count || total > 0 {
 		if err := tx.Session(&gorm.Session{}).
-			Offset(int((p.Page - 1) * p.PageSize)).
+			Offset((p.Page - 1) * p.PageSize).
 			Limit(p.PageSize).
 			Find(&list).Error; err != nil {
 			return nil, err
